refactor(carousel) PR changes
