--- app/components/Carousel/components/PogressBarItem/index.ts
@@ -1,1 +1,0 @@
-export { PogressBarItem } from './PogressBarItem';
