+++ app/components/Carousel/components/ProgressBarItem/ProgressBarItem.data.ts
