+++ app/components/Carousel/components/ProgressBarItem/ProgressBarItem.props.ts
@@ -1,6 +1,6 @@
 import Animated from 'react-native-reanimated';
 
-export interface IPogressBarItemProps {
+export interface IProgressBarItemProps {
   index: number;
   length: number;
   animValue: Animated.SharedValue<number>;
