+++ app/components/Carousel/components/ProgressBarItem/index.ts
@@ -1,0 +1,1 @@
+export { ProgressBarItem } from './ProgressBarItem';
