+++ app/components/Carousel/components/index.ts
@@ -1,3 +1,3 @@
 export * from './ImageSubTitle/ImageSubTitle';
 export * from './ImageTitleSubTitle';
-export * from './PogressBarItem';
+export * from './ProgressBarItem';
